"Changed button margin and added register button in index.html"
This commit message describes the changes made in the code diff for the file index.html. It mentions that the margin top of the button has been changed from 50px to 30px, and a new button for registering has been added. The message is concise and accurately describes the changes made in the code.