Delete HTML files for connection and registration and PHP function file
This commit deletes the following files:

* index.html
* v1.0/HTML/register.html
* v1.0/PHP/function.php

The deleted files are related to the connection and registration functionality of an application. The HTML files contain the front-end interface for user interaction, while the PHP file contains back-end functions for processing user data. The reason for deleting these files is not clear from the provided context.